adding fx effect, spawning defenders on click and ui to select defenders

+++ Assets/Scripts/LevelLoader.cs
@@ -5,7 +5,7 @@ using UnityEngine.SceneManagement;
 
 public class LevelLoader : MonoBehaviour
 {
-    [SerializeField] float timeToWait = 4;
+    [SerializeField] int timeToWait = 4;
 
     int currentSceneIndex;
